Test: Update prompt creation test with heading at top
gptel-create-prompt-test.el (gptel-test-prompt-branching-heading-at-top):
Add test for branching context in org with a heading
at (point-min).

The heading should be included if it's part of the current lineage
and not be included if it's not.  At the same time non-heading
text at (point-min) is always part of the lineage and should be included.